Player Object Referencing
The player object is now assigned correctly on the client side and not just server side. This means commands can now be issued to the player controller from the player connection object. This is important for interactions between players and other scripts. The example this is used for is the flashlight toggle, which has now been implemented.