Fixed #5
Fixed a bug where data values were not being synced correctly when they were constantly changing. This is now fixed. The cursor is now enabled when the player leaves the game, for whatever reason, to make sure that they can actually use the buttons to reconnect or any other action.