A* Walkable Pathfinding
Improved A* pathfinding to check for walkable areas as the algorithm works, instead of having a precompiled area where AI cannot travel. This means the environment can change, e.g. buildings placed and AI will still work around it. Also added settings to AI testing script, in order to change pathfinding grid size, debug mode or change the check radius for walkable areas. Made it so certain tags are marked as unwalkable. This will be useful for things such as buildings/bunker parts.