Fixed a few location issues with AI
Fixed a few issues relating to the Z axis being incorrectly marked as the Y axis when usingh Vector3's within AI code. Also added grid origin position to position.transform so the object that has the grid script on it can be moved about without needing to use the origin position vector.