Created AI Chase Mode
AI now have the ability to chase a target to a certain position. Corridors also now have a building tag associated with them in order for them to be identified by AI as a potential target for attack. The player layer has also been renamed to Entity in order to make it so all entities (player, AI) no longer collide as this would cause pathfinding issues if they did.